+++ src/brasil/gov/tiles/tiles/collection.py
@@ -6,8 +6,9 @@ from collective.cover.tiles.base import PersistentCoverTile
 from collective.cover.tiles.configuration_view import IDefaultConfigureForm
 from plone import api
 from plone.app.uuid.utils import uuidToObject
-from plone.directives import form
+from plone.autoform import directives as form
 from plone.namedfile.field import NamedBlobImage as NamedImage
+from plone.supermodel import model
 from plone.tiles.interfaces import ITileDataManager
 from plone.tiles.interfaces import ITileType
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
@@ -16,7 +17,7 @@ from zope.component import queryUtility
 from zope.schema import getFieldsInOrder
 
 
-class ICollectionTile(IPersistentCoverTile, form.Schema):
+class ICollectionTile(IPersistentCoverTile, model.Schema):
 
     header = schema.TextLine(
         title=_(u'Header'),