+++ CHANGES.rst
@@ -4,7 +4,7 @@ Alterações
 1.4 (unreleased)
 ^^^^^^^^^^^^^^^^
 
-- Nothing changed yet.
+- Remove 'five.grok' e plone.directives.form (closes `#185`)
 
 
 1.4rc1 (2017-10-05)